requirements.txt
dask==2023.5.0
datasets==2.14.6
datasketch==1.6.4
datawig==0.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dgl==1.1.1
dill==0.3.7
duckdb==0.9.2
ecos==2.0.12
elastic-transport==8.10.0
elasticsearch==8.10.1
exceptiongroup==1.2.0
executing==2.0.0
faiss-gpu==1.7.2
fancyimpute==0.7.0
networkx==3.0rc1
pyarrow==13.0.0
pytrec-eval==0.5
pyspark==3.5.0
scikit-learn==1.3.2
scipy==1.10.1
tabulate==0.9.0
torch==1.12.1+cu102
torch-scatter==2.1.2
torch-sparse==0.6.18
torch_geometric==2.4.0
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
transformers==4.33.1
transtab==0.0.5